check if messages directory exists

+++ clean-tenants.js
@@ -11,11 +11,12 @@ tenantDirs.forEach(tenantDir => {
     }
 
     const messagesDir = `${tenantDir}/messages`;
+    if (fs.existsSync(messagesDir)) {
+        const messagesFilesToRemove = fs.readdirSync(messagesDir).filter(file => file.endsWith('-messages.json') && file !== 'all-messages.json');
+        messagesFilesToRemove.forEach(file => fs.unlinkSync(`${messagesDir}/${file}`));
 
-    const messagesFilesToRemove = fs.readdirSync(messagesDir).filter(file => file.endsWith('-messages.json') && file !== 'all-messages.json');
-    messagesFilesToRemove.forEach(file => fs.unlinkSync(`${messagesDir}/${file}`));
-
-    const messagesFilesToRename = fs.readdirSync(messagesDir).filter(file => file.startsWith('2021-') || file.startsWith('2022-'));
-    messagesFilesToRename.forEach(file => fs.renameSync(`${messagesDir}/${file}`, `${messagesDir}/${file.slice(21)}`));
+        const messagesFilesToRename = fs.readdirSync(messagesDir).filter(file => file.startsWith('2021-') || file.startsWith('2022-'));
+        messagesFilesToRename.forEach(file => fs.renameSync(`${messagesDir}/${file}`, `${messagesDir}/${file.slice(21)}`));
+    }
 
 });
